first test for RxBuilder
